- make vespalib::strivg construction from std::string_view explicit to bring it closer to std::string. - Clean up mix and matcg with string/string_view.

+++ searchlib/src/vespa/searchlib/queryeval/sourceblendersearch.cpp
@@ -148,7 +148,7 @@ SourceBlenderSearch::visitMembers(vespalib::ObjectVisitor &visitor) const
     visit(visitor, "children", _children);
     for (const auto & child : _children) {
         vespalib::asciistream os;
-        os << "Source " << child;
+        os << "Source " << uint32_t(child);
         visit(visitor, os.str(), *getSearch(child));
     }
 }
